+++ lib/pagy/countless.rb
@@ -11,6 +11,7 @@ class Pagy
       normalize_vars(vars)
       setup_vars(page: 1, outset: 0)
       setup_items_var
+      setup_params_var
       @offset = (@items * (@page - 1)) + @outset
     end