+++ test/pagy/extras/standalone_test.rb
@@ -74,6 +74,17 @@ describe 'pagy/extras/standalone' do
       _(app.pagy_url_for(pagy, 5)).must_equal "?a&page=5#fragment"
       _(app.pagy_url_for(pagy, 5, absolute: true)).must_equal "?a&page=5#fragment"
     end
+    it 'renders url with params lambda and fragment' do
+      pagy = Pagy.new count: 1000, page: 3, params: ->(p) { p.merge(a: 3, b: 4) }, fragment: '#fragment'
+      _(app.pagy_url_for(pagy, 5)).must_equal '/foo?page=5&a=3&b=4#fragment'
+      _(app.pagy_url_for(pagy, 5, absolute: true)).must_equal 'http://example.com:3000/foo?page=5&a=3&b=4#fragment'
+      pagy = Pagy.new count: 1000, page: 3, params: ->(p) { p.merge(a: [1, 2, 3]) }, fragment: '#fragment', url: 'http://www.pagy-standalone.com/subdir'
+      _(app.pagy_url_for(pagy, 5)).must_equal "http://www.pagy-standalone.com/subdir?page=5&a[]=1&a[]=2&a[]=3#fragment"
+      _(app.pagy_url_for(pagy, 5, absolute: true)).must_equal "http://www.pagy-standalone.com/subdir?page=5&a[]=1&a[]=2&a[]=3#fragment"
+      pagy = Pagy.new count: 1000, page: 3, params: ->(p) { p.merge(a: nil) }, fragment: '#fragment', url: ''
+      _(app.pagy_url_for(pagy, 5)).must_equal "?page=5&a#fragment"
+      _(app.pagy_url_for(pagy, 5, absolute: true)).must_equal "?page=5&a#fragment"
+    end
   end
 end